helium/ui: fix caption bounds (only on macos for now)
merged the related patch from macOS repo into this one and added new changes. it makes sense to keep it here as it's a "main helium" kind of change, not some platform quirk.

i'm not sure whether macOS 15 bounds are correct, however, so this will probably be patched up more in the future.